合并搜索路由
